// client/src/util/store/action/actionType.tsx
export enum actionType {
    SET_ACTIVE_SIDEBAR= "SET_ACTIVE_SIDEBAR",
    SET_SCREEN_SIZE= "SET_SCREEN_SIZE",
    SET_LOADING_COLLECTION="SET_LOADING_COLLECTION",
    SET_LOADING_SWIPER="SET_LOADING_SWIPER",
    SET_LOADING_DETAIL="SET_LOADING_DETAIL",
    SET_GOOGLE_STATUS="SET_GOOGLE_STATUS",
    SET_AXIOS_MESSAGE="SET_AXIOS_MESSAGE",
    SET_DATA_COLLECTION = "SET_DATA_COLLECTION",
    SET_DATA_SWIPER = "SET_DATA_SWIPER",
    SET_ADD_NEW = "SET_ADD_NEW",
    SET_REQUEST_LOADING = "SET_REQUEST_LOADING",
    SET_OPEN_CONNECT = "SET_OPEN_CONNECT",
    SET_ACCOUNT_INFO = "SET_ACCOUNT_INFO",
    SET_COMPARE_DATA = "SET_COMPARE_DATA",
    SET_YOUR_COLLECTION = "SET_YOUR_COLLECTION",
    SET_OPEN_COMPARE = "SET_OPEN_COMPARE",
    SET_ITEM_REVIEW = "SET_ITEM_REVIEW",
    SET_LOADING_YOUR_COLLECTION = "SET_LOADING_YOUR_COLLECTION",
    SET_SHOPPING_CART = "SET_SHOPPING_CART"
}